requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
websockets==12.0
pyyaml==6.0.1
pyjwt==2.8.0
aiohttp==3.9.1
livekit==0.9.1
cartesia==1.0.0
deepgram-sdk==3.0.0
python-dotenv==1.0.0